#198 [REFACTOR] utils 폴더 아래 js 파일을 default export가 아닌 일반 export로 변경
